Update legend-item segmentation
Update legend-item segmentation to add internal json output and bypass mapKurator outputs.

+++ segmentation/legend_item_segmentation/link_inference.py
@@ -408,18 +408,21 @@ def read_results_from_mapkurator(target_map_name, input_image, path_to_intermedi
 
     map_name = target_map_name
     #mapkurator_source_name = os.path.join(path_to_mapkurator_output, map_name.replace('.tif', '.geojson'))
-    mapkurator_source_name = path_to_mapkurator_output
-    mapkurator_name = os.path.join(path_to_intermediate, 'intermediate3_2', map_name.replace('.tif', '_v2.geojson'))
-    basemap_name = os.path.join(path_to_intermediate, 'area_crop_rgb.tif')
+    if os.path.isfile(path_to_mapkurator_output) == True:
+        mapkurator_source_name = path_to_mapkurator_output
+        mapkurator_name = os.path.join(path_to_intermediate, 'intermediate3_2', map_name.replace('.tif', '_v2.geojson'))
 
+        with open(mapkurator_source_name, 'r') as file:
+            source_data = file.read().replace('-', '')
+        with open(mapkurator_name, 'w') as file:
+            file.write(source_data)
+    else:
+        mapkurator_name = os.path.join(path_to_intermediate, 'intermediate9', map_name.replace('.tif', '_tesseract_v2.geojson'))
+    
+    
+    basemap_name = os.path.join(path_to_intermediate, 'area_crop_rgb.tif')
     base_image = cv2.imread(basemap_name)
     base_image = cv2.cvtColor(base_image, cv2.COLOR_BGR2GRAY)
-
-
-    with open(mapkurator_source_name, 'r') as file:
-        source_data = file.read().replace('-', '')
-    with open(mapkurator_name, 'w') as file:
-        file.write(source_data)
 
     gdf = gpd.read_file(mapkurator_name, driver='GeoJSON')
     text_mask = np.zeros((base_image.shape[0], base_image.shape[1]), dtype='uint8')
@@ -1665,6 +1668,70 @@ def generating_geojson(target_map_name, input_image, path_to_intermediate):
 
 
 
+def generating_json(target_map_name, input_image, output_dir, path_to_intermediate):
+    print('Step (9/9): Preparing output json - Generating JSON file (competition format)...')
+
+    map_name = target_map_name
+
+    linked_poly_description1 = gpd.read_file(os.path.join(path_to_intermediate, 'intermediate7', map_name.replace('.tif', '_PolygonType.geojson')), driver='GeoJSON')
+    new_item = []
+
+    for index, row in linked_poly_description1.iterrows():
+        if len(row['abbreviation']) > 0:
+            this_poly_name = row['abbreviation'] + '_poly'
+        else:
+            this_poly_name = str(row['id']) + '_poly'
+        this_legend_bounds = row['geometry'].bounds
+        this_item = {}
+        this_item['label'] = this_poly_name
+        this_item['points'] = [[this_legend_bounds[0], -1.0*this_legend_bounds[1]], [this_legend_bounds[2], -1.0*this_legend_bounds[3]]]
+        this_item['group_id'] = None
+        this_item['shape_type'] = 'rectangle'
+        this_item['flags']: None
+        new_item.append(this_item)
+
+    updated_record = {}
+    updated_record['version'] = '0.0.1'
+    updated_record['flags'] = None
+    updated_record['shapes'] = new_item
+
+    with open(os.path.join(output_dir, map_name.replace('.tif', '_PolygonType_internal.json')), 'w') as outfile: 
+        json.dump(updated_record, outfile)
+
+
+
+
+    linked_poly_description1 = gpd.read_file(os.path.join(path_to_intermediate, 'intermediate7', map_name.replace('.tif', '_PointLineType.geojson')), driver='GeoJSON')
+    new_item = []
+
+    for index, row in linked_poly_description1.iterrows():
+        if len(row['abbreviation']) > 0:
+            this_poly_name = row['abbreviation'] + '_poly'
+        else:
+            this_poly_name = str(row['id']) + '_poly'
+        this_legend_bounds = row['geometry'].bounds
+        this_item = {}
+        this_item['label'] = this_poly_name
+        this_item['points'] = [[this_legend_bounds[0], -1.0*this_legend_bounds[1]], [this_legend_bounds[2], -1.0*this_legend_bounds[3]]]
+        this_item['group_id'] = None
+        this_item['shape_type'] = 'rectangle'
+        this_item['flags']: None
+        new_item.append(this_item)
+
+    updated_record = {}
+    updated_record['version'] = '0.0.1'
+    updated_record['flags'] = None
+    updated_record['shapes'] = new_item
+
+    with open(os.path.join(output_dir, map_name.replace('.tif', '_PointLineType_internal.json')), 'w') as outfile: 
+        json.dump(updated_record, outfile)
+
+
+
+    return True
+
+
+
 
 def adjusting_crs(target_map_name, input_image, path_to_intermediate, output_dir, postprocessing_for_crs):
     if not os.path.exists(output_dir):
@@ -1738,6 +1805,7 @@ def start_linking(target_map_name, input_image, output_dir, path_to_intermediate
     linking_description_pointline(target_map_name, input_image, path_to_intermediate)
     integrating_result(target_map_name, input_image, path_to_intermediate)
     generating_geojson(target_map_name, input_image, path_to_intermediate)
+    generating_json(target_map_name, input_image, output_dir, path_to_intermediate)
     adjusting_crs(target_map_name, input_image, path_to_intermediate, output_dir, postprocessing_for_crs)
